requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
flake8==6.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.5.0
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.1.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
